pyproject.toml
[project]
name = "memevolve"
version = "0.1.0"
description = "MemEvolve: Adaptive Memory System for AI Agents"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "MemEvolve Contributors"}
]
keywords = ["memory", "ai", "adaptive", "evolution", "llm"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "numpy>=1.24.0",
    "openai>=2.15.0",
    "faiss-cpu>=1.7.4",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "requests>=2.31.0",
    "tiktoken>=0.12.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-timeout>=2.2.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-timeout>=2.2.0",
    "flake8>=7.3.0",
    "autopep8>=2.3.0",
]
neo4j = [
    "neo4j>=5.0.0",
]
gpu = [
    "faiss-gpu>=1.7.4",
]

[project.urls]
Homepage = "https://github.com/thephimart/MemEvolve-API"
Repository = "https://github.com/thephimart/MemEvolve-API"
Issues = "https://github.com/thephimart/MemEvolve-API/issues"

[project.scripts]
memevolve-server = "memevolve.api.server:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
memevolve = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
timeout = 600
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v"

[tool.flake8]
max-line-length = 100
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    ".eggs",
    "*.egg",
]

[tool.autopep8]
max_line_length = 100
aggressive = 2
in-place = true
recursive = true